fix: CUDA 12.x compatibility for Pascal GPU and improve test script
- Install CUDA 12.6 alongside CUDA 13.1 (CUDA 13.1 incompatible with Pascal GPUs)
- Update test script to prefer CUDA 12.x for Pascal GPU compatibility (compute 6.1)
- Fix OpenMP loop variables: size_t -> long long for MSVC compatibility
- Add NOMINMAX define to resource_monitor.cpp for Windows min/max macro conflicts
- Add cuComplex.h include to distributed_gpu.cu for CUDA 12.6 compatibility
- Fix linker error: run_lret_simulation -> run_simulation_optimized in gpu_simulator.cu
- Update CMakeLists.txt: add compute_61 to CUDA_ARCHITECTURES for Pascal support
- Improve run_hardware_dependent_tests.py:
  - find_cuda_path() now prefers CUDA 12.x versions when available
  - setup_windows_build_environment() accepts prefer_cuda_12x parameter
  - get_cmake_generator_and_env() passes CUDA preference to environment setup
  - Add CLI arguments: --cuda-12x, --cuda-13x, --use-existing-build
  - Fix executable detection: strip ./ and add .exe extension on Windows
  - Add actual_build_path parameter to run_single_test() for using existing builds

Results:
- All 19 executables built successfully with CUDA 12.6
- test_distributed_gpu passes on single Quadro P400 GPU (0.19s)
- Multi-GPU tests appropriately skipped (insufficient hardware)

+++ src/gate_fusion.cpp
@@ -323,7 +323,7 @@ MatrixXcd apply_fused_gate(
     #ifdef USE_OPENMP
     #pragma omp parallel for if(dim > 256)
     #endif
-    for (size_t i = 0; i < dim / 2; ++i) {
+    for (long long i = 0; i < (long long)(dim / 2); ++i) {
         // Calculate the two indices that differ in the target qubit
         size_t i0 = (i & ~(mask - 1)) << 1 | (i & (mask - 1));  // target qubit = 0
         size_t i1 = i0 | mask;  // target qubit = 1
