+++ src/simd_kernels.cpp
@@ -101,7 +101,7 @@ MatrixXcd apply_single_qubit_core(
 #ifdef _OPENMP
     #pragma omp parallel for schedule(static) if(dim > 4096 && rank > 2)
 #endif
-    for (size_t block = 0; block < dim; block += 2 * step) {
+    for (long long block = 0; block < (long long)dim; block += 2 * step) {
         for (size_t i = block; i < block + step && i < dim; ++i) {
             size_t i0 = i;
             size_t i1 = i + step;
